requirements.txt
Flask==3.0.0
mygeotab==0.8.5
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Werkzeug==3.0.1
